+++ example/lib/main.dart
@@ -8,6 +8,7 @@ import 'screens/chinese_screen.dart';
 import 'screens/custom_theme_screen.dart';
 import 'screens/data_backed_calendar_screen.dart';
 import 'screens/date_time_screen.dart';
+import 'screens/dropin_screen.dart';
 import 'screens/hijri_screen.dart';
 import 'screens/inline_form_screen.dart';
 import 'screens/localization_screen.dart';
@@ -58,6 +59,7 @@ class HomeScreen extends StatelessWidget {
     ('Time only', '12-hour and 24-hour time picker', _time),
     ('Hijri', 'Umm al-Qura calendar in Arabic', _hijri),
     ('Chinese', 'Lunisolar calendar with leap months', _chinese),
+    ('Drop in', 'Cupertino and Calendar replacements', _dropin),
     ('Data backed', 'Committee calendar from a dataset', _dataBacked),
     ('Basic', 'Drop-in replacement for showDatePicker', _basic),
     ('Birth date', 'Drum mode, locked, min/max age', _birth),
@@ -74,6 +76,7 @@ class HomeScreen extends StatelessWidget {
   static Widget _time(BuildContext c) => const TimeScreen();
   static Widget _hijri(BuildContext c) => const HijriScreen();
   static Widget _chinese(BuildContext c) => const ChineseScreen();
+  static Widget _dropin(BuildContext c) => const DropinScreen();
   static Widget _dataBacked(BuildContext c) => const DataBackedCalendarScreen();
   static Widget _basic(BuildContext c) => const BasicScreen();
   static Widget _birth(BuildContext c) => const BirthDateScreen();